requirements.txt
anthropic
python-dotenv
tweepy
schedule
pandas